Patch from Hans Feldt This small patch (against r948) adds variable CPPFLAGS to all make rules that compiles `.c' files into `.o' files.
This makes it possible to set CPPFLAGS on the command line as in:

$ make CPPFLAGS=-DAMFDEBUG

Editing the makefile when prototyping is then not needed.

Another reason is that some of the openais C files are built using the 
implicit rule for C files (which uses CPPFLAGS). The change makes rules 
consistent with the implicit rule.

It also adds a few missing files to the clean make target.



git-svn-id: https://svn.fedorahosted.org/svn/openais/trunk@950 fd59a12c-fef9-0310-b244-a6a79926bd2f