Add a configuraiton option to allow generated nodeids to be limited to 31-bits
Some openais clients require a signed 32 bit nodeid that is greater
than zero however by default openais uses all 32 bits of the IPv4 address
space when generating a nodeid.  Setting this option to yes ensures any
generated nodeid is a positive signed 32 bit integer.

The clusters behavior is undefined if this option is enabled on only
a subset of the cluster (for example during a rolling upgrade) and should
be enabled with caution.



git-svn-id: https://svn.fedorahosted.org/svn/openais/branches/whitetank@1655 fd59a12c-fef9-0310-b244-a6a79926bd2f