Correctly compare local addresses with bindnetaddr
This patch allows whitetank/corosync to behave correctly when the supplied value of bindnetaddr has more precision than the netmask, but is not a full IPv4 address.

For example, currently, if the netmask is /16 but the user specifies 192.168.1.0, then OpenAIS is unable to determine the correct address and uses the loopback address instead.

The following patch allows OpenAIS to correctly determine which interface/address to use.


git-svn-id: https://svn.fedorahosted.org/svn/openais/branches/whitetank@1665 fd59a12c-fef9-0310-b244-a6a79926bd2f