Backport of Corosync 6aa47fde953bf2179f5bd2dd07815fc7d80f47bb
A commit token should be rejected when a token is lost in the recovery
state.  This occurs naturally because the ring id increases by 4 for
every new ring.  Prior to this patch, if the token was lost, the old
ring id information was restored, causing a commit token to be accepted
when it should be rejected.  This erronously accepted commit token would
lead to an assertion which is fixed by this patch.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>
Reviewed-by: Steven Dake <sdake@redhat.com>



git-svn-id: https://svn.fedorahosted.org/svn/openais/branches/whitetank@2165 fd59a12c-fef9-0310-b244-a6a79926bd2f