totemsrp: Remove recv_flush code
Backport of Corosync 336741ee96caf3ae435b609ee8a76788902c9edf

The recv_flush code is no longer necessary because of the miss_count_count
addition.  It can in some cases lead to register corruption because of
interactions with -fstack-protector, the recursive nature of how this code
works, and interactions with the optimizer in some versions of gcc.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>
Reviewed-by: Steven Dake <sdake@redhat.com>


git-svn-id: https://svn.fedorahosted.org/svn/openais/branches/whitetank@2167 fd59a12c-fef9-0310-b244-a6a79926bd2f