totemsrp: Only restore old ring id information one time
Backport of Corosync 0eabeee63eca7a4cc1d907607057ac668fafbcae

The current code stores the current ring information every time a commit
token is generated.  This causes the old ring id used for comparison purposes
to increase if a token is lost in commit or recovery, resulting in failure of
totem.  This patch changes the behavior to only store the old ring id one
time when the commit token is received, and then further commit token ring
id saves are not done until OPERATIONAL is reached.

Signed-off-by: Jan Friesse <jfriesse@redhat.com>
Reviewed-by: Steven Dake <sdake@redhat.com>


git-svn-id: https://svn.fedorahosted.org/svn/openais/branches/whitetank@2168 fd59a12c-fef9-0310-b244-a6a79926bd2f