Fix coredump when rolling upgrade to Corosync 1.3
This patch fixes the rolling upgrade issue from openais whitetank to
corosync 1.3. The problem is the length of the array
cpg_exec_service[]/cpg_exec_engine[] is different in the two version,
where in corosync is one more service.

So the "old" node (with openais whitetank) will get a NULL pointer
when it recevies the message from the "new" node (with corosync 1.3)
and then coredump.

Signed-off-by: Jiaju Zhang <jjzhang@suse.de>
Reviewed-by: Steven Dake <sdake@redhat.com>


git-svn-id: https://svn.fedorahosted.org/svn/openais/branches/whitetank@2174 fd59a12c-fef9-0310-b244-a6a79926bd2f