whitetank ipc: Don't deadlock in ipc_disconnect
If ipc_disconnect was called in msg_send and this was called by
outq_flush, there was deadlock if send failed on error different
then EINTR and EAGAIN. This problem is now solved by adding
ipc_disconnect_locked function which takes parameter locked and
depending on it locks mutex or not. Old function ipc_disconnect
is converted to call of new function by using locked parameter
set to 0.                                                                                                                   
Signed-off-by: Jan Friesse <jfriesse@redhat.com>
Reviewed-by: Steven Dake <sdake@redhat.com>



git-svn-id: https://svn.fedorahosted.org/svn/openais/branches/whitetank@2180 fd59a12c-fef9-0310-b244-a6a79926bd2f