Fix bug when start/stop OpenAIS repeatedly
The root cause of this issue is if two membership changes are too
close, before finishing the execution of all the callbacks initiated
by the first membership change, the second membership change arrived.
Since some internal status of those callbacks has not been reset to
initial state, re-enter those callbacks will cause problems.

The problematic callback here is the ckpt_confchg_fn in openais CKPT
service. The mentioned internal status is my_sync_state. So, before
my_sync_state goes back to SYNC_STATE_NOT_STARTED, another membership
change happens, this results in the new callback is wrongly rejected
and will never execute again. Thus, that CKPT cannot be sycned
normally and end up with refcounting of a particular checkpoint is not
correct in the cluster so being deleted then. This is why the user of
CKPT service - ocfs2_controld sees the "Object does not exist", but it
should exist. The solution in that patch is to judge the ringid first.
That being said, if the ringid is not the same, that indicates it is a
new membership change, so it is not re-entering an ongoing
ckpt_confchg_fn, but needs to start a new ckpt_confchg_fn indeed.

The second part in that patch which needs to be mentioned here is the
changing of when to assign value to my_saved_ring_id. When the totem
configuration type is still TOTEM_CONFIGURATION_TRANSITIONAL, it seems
to be not very precise if assigning a new ringid value to
my_saved_ring_id, since the old configuration change has not been
finished yet.

Signed-off-by: Lidong Zhong <lzhong@suse.com>
Patch-description-by: Jiaju Zhang <jjzhang@suse.de>
Reviewed-by: Jan Friesse <jfriesse@redhat.com>


git-svn-id: https://svn.fedorahosted.org/svn/openais/trunk@2190 fd59a12c-fef9-0310-b244-a6a79926bd2f